chore: move aliased/deprecated operators into separate files (#6080)
* chore: move exhaust into separate file

* chore: move combineLatest into separate file

* chore: move concat into separate file

* chore: move merge into separate file

* chore: move flatMap into separate file

* chore: move race into a separate file

* chore: move zip into separate file
