+++ src/internal/operators/exhaust.ts
@@ -1,0 +1,6 @@
+import { exhaustAll } from './exhaustAll';
+
+/**
+ * @deprecated renamed. Use {@link exhaustAll}.
+ */
+export const exhaust = exhaustAll;
