+++ src/internal/operators/exhaustAll.ts
@@ -75,8 +75,3 @@ export function exhaustAll<T>(): OperatorFunction<any, T> {
     );
   });
 }
-
-/**
- * @deprecated renamed. Use {@link exhaustAll}.
- */
-export const exhaust = exhaustAll;
