+++ src/internal/operators/flatMap.ts
@@ -1,0 +1,6 @@
+import { mergeMap } from './mergeMap';
+
+/**
+ * @deprecated renamed. Use {@link mergeMap}.
+ */
+export const flatMap = mergeMap;
