+++ src/operators/index.ts
@@ -9,11 +9,13 @@ export { bufferWhen } from '../internal/operators/bufferWhen';
 export { catchError } from '../internal/operators/catchError';
 export { combineAll } from '../internal/operators/combineAll';
 export { combineLatestAll } from '../internal/operators/combineLatestAll';
-export { combineLatest, combineLatestWith } from '../internal/operators/combineLatestWith';
+export { combineLatest } from '../internal/operators/combineLatest';
+export { combineLatestWith } from '../internal/operators/combineLatestWith';
+export { concat } from '../internal/operators/concat';
 export { concatAll } from '../internal/operators/concatAll';
 export { concatMap } from '../internal/operators/concatMap';
 export { concatMapTo } from '../internal/operators/concatMapTo';
-export { concat, concatWith } from '../internal/operators/concatWith';
+export { concatWith } from '../internal/operators/concatWith';
 export { connect } from '../internal/operators/connect';
 export { count } from '../internal/operators/count';
 export { debounce } from '../internal/operators/debounce';
@@ -28,7 +30,8 @@ export { distinctUntilKeyChanged } from '../internal/operators/distinctUntilKeyC
 export { elementAt } from '../internal/operators/elementAt';
 export { endWith } from '../internal/operators/endWith';
 export { every } from '../internal/operators/every';
-export { exhaust, exhaustAll } from '../internal/operators/exhaustAll';
+export { exhaust } from '../internal/operators/exhaust';
+export { exhaustAll } from '../internal/operators/exhaustAll';
 export { exhaustMap } from '../internal/operators/exhaustMap';
 export { expand } from '../internal/operators/expand';
 export { filter } from '../internal/operators/filter';
@@ -44,11 +47,13 @@ export { map } from '../internal/operators/map';
 export { mapTo } from '../internal/operators/mapTo';
 export { materialize } from '../internal/operators/materialize';
 export { max } from '../internal/operators/max';
-export { mergeWith, merge } from '../internal/operators/mergeWith';
+export { merge } from '../internal/operators/merge';
 export { mergeAll } from '../internal/operators/mergeAll';
-export { mergeMap, flatMap } from '../internal/operators/mergeMap';
+export { flatMap } from '../internal/operators/flatMap';
+export { mergeMap } from '../internal/operators/mergeMap';
 export { mergeMapTo } from '../internal/operators/mergeMapTo';
 export { mergeScan } from '../internal/operators/mergeScan';
+export { mergeWith } from '../internal/operators/mergeWith';
 export { min } from '../internal/operators/min';
 export { multicast } from '../internal/operators/multicast';
 export { observeOn } from '../internal/operators/observeOn';
@@ -60,7 +65,8 @@ export { publish } from '../internal/operators/publish';
 export { publishBehavior } from '../internal/operators/publishBehavior';
 export { publishLast } from '../internal/operators/publishLast';
 export { publishReplay } from '../internal/operators/publishReplay';
-export { race, raceWith } from '../internal/operators/raceWith';
+export { race } from '../internal/operators/race';
+export { raceWith } from '../internal/operators/raceWith';
 export { reduce } from '../internal/operators/reduce';
 export { repeat } from '../internal/operators/repeat';
 export { repeatWhen } from '../internal/operators/repeatWhen';
@@ -103,5 +109,6 @@ export { windowTime } from '../internal/operators/windowTime';
 export { windowToggle } from '../internal/operators/windowToggle';
 export { windowWhen } from '../internal/operators/windowWhen';
 export { withLatestFrom } from '../internal/operators/withLatestFrom';
-export { zip, zipWith } from '../internal/operators/zipWith';
+export { zip } from '../internal/operators/zip';
 export { zipAll } from '../internal/operators/zipAll';
+export { zipWith } from '../internal/operators/zipWith';
